+++ testo.php
@@ -2,6 +2,7 @@
 
 declare(strict_types=1);
 
+use Rasuvaeff\Understudy\Testo\UnderstudyPlugin;
 use Testo\Application\Config\ApplicationConfig;
 use Testo\Application\Config\FinderConfig;
 use Testo\Application\Config\SuiteConfig;
@@ -12,6 +13,7 @@ return new ApplicationConfig(
         new SuiteConfig(
             name: 'Unit',
             location: new FinderConfig(include: ['tests'], exclude: ['tests/Integration']),
+            plugins: [new UnderstudyPlugin()],
         ),
         new SuiteConfig(
             name: 'Integration',